# cmd/docs/swagger.yaml
definitions:
...
  dto.SaleInvoiceRequestDTO:
    properties:
      customerId:
        type: integer
      discount:
        type: integer
      grandTotal:
        type: integer
      id:
        type: string
      remark:
        type: string
      saleDate:
        type: string
      saleDetails:
        items:
          $ref: '#/definitions/models.SaleDetail'
        type: array
      total:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  httputil.HttpError400:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  httputil.HttpError401:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized
        type: string
    type: object
  httputil.HttpError500:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: internal server error
        type: string
    type: object
...
  models.Customer:
    properties:
      address:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updatedAt:
        type: string
    type: object
  models.Product:
    properties:
      brand:
        type: string
      buyPrice:
        minimum: 1
        type: integer
      categoryId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      isActive:
        type: boolean
      productName:
        minLength: 3
        type: string
      qtyOhHand:
        type: integer
      reorderlvl:
        minimum: 1
        type: integer
      sellPriceLevel1:
        minimum: 1
        type: integer
      sellPriceLevel2:
        minimum: 1
        type: integer
      uom:
        minLength: 3
        type: string
      updatedAt:
        type: string
    required:
    - buyPrice
    - productName
    - qtyOhHand
    - reorderlvl
    - sellPriceLevel1
    - sellPriceLevel2
    - uom
    type: object
...
  models.Sale:
    properties:
      createdAt:
        type: string
      customer:
        $ref: '#/definitions/models.Customer'
      customerId:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      discount:
        type: integer
      grandTotal:
        type: integer
      id:
        type: string
      remark:
        type: string
      saleDate:
        type: string
      saleDetails:
        items:
          $ref: '#/definitions/models.SaleDetail'
        type: array
      total:
        type: integer
      updatedAt:
        type: string
    type: object
  models.SaleDetail:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      price:
        type: integer
      productId:
        type: string
      productName:
        type: string
      qty:
        type: integer
      saleId:
        type: string
      total:
        type: integer
      updatedAt:
        type: string
    type: object
...
info:
  contact:
    email: sankaungkin@gmail.com
    name: API Support
  description: This is an auto-generated API docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber API
  version: "1.0"
paths:
...
  /api/sale:
    get:
      consumes:
      - application/json
      description: Fetch all sales
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Sale'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - ApiKeyAuth: []
      - Bearer  <-----------------------------------------add this in all controllers that need authentication: []
      summary: Fetch all sales
      tags:
      - Sales
    post:
      consumes:
      - application/json
      description: Create new sale invoice based on parameters
      parameters:
      - description: Sale Data
        in: body
        name: sale
        required: true
        schema:
          $ref: '#/definitions/dto.SaleInvoiceRequestDTO'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sale'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - ApiKeyAuth: []
      - Bearer  <-----------------------------------------add this in all controllers that need authentication: []
      summary: Create new sale invoice based on parameters
      tags:
      - Sales
  /api/sale/{id}:
    get:
      consumes:
      - application/json
      description: Fetch individual sale invoice by Id
      parameters:
      - description: sale Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - ApiKeyAuth: []
      - Bearer  <-----------------------------------------add this in all controllers that need authentication: []
      summary: Fetch individual sale invoice by Id
      tags:
      - Sales
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"